fix!: Handle long verbosity flag correctly
Before, `--verbosity` worked just like `-v`: it could be repeatedly
added to increase logging output. Also, `-v` was configured in a way
that passing it once actually changed nothing.

Now, the long flag works to _set_ the verbosity:

    --verbosity=0
    --verbosity 4

And the short flag can still be repeated up to three times.